Fix the mouse selection up-scrolling: fontHeight is an unsigned int, m_scrollY an int, now in the division, the int is casted to unsigned which isn't very nice when we want to scroll upwards (negative m_scrollY).
renderer()->fontHeight() is used quite often, and I wonder if it wouldn't be
safer to make it return an int to avoid more mistakes..  what do you guys think?

CCBUG: 151795

svn path=/trunk/KDE/kdelibs/kate/; revision=733222

+++ part/view/kateviewinternal.cpp
@@ -2799,7 +2799,7 @@ void KateViewInternal::scrollTimeout ()
 {
   if (m_scrollX || m_scrollY)
   {
-    scrollLines (startPos().line() + (m_scrollY / renderer()->fontHeight()));
+    scrollLines (startPos().line() + (m_scrollY / (int) renderer()->fontHeight()));
     placeCursor( QPoint( m_mouseX, m_mouseY ), true );
   }
 }
